+++ lib/galaxy/managers/collections.py
@@ -34,10 +34,6 @@ from galaxy.managers.hdas import (
 from galaxy.managers.hdcas import write_dataset_collection
 from galaxy.managers.histories import HistoryManager
 from galaxy.managers.lddas import LDDAManager
-from galaxy.managers.short_term_storage import (
-    ShortTermStorageMonitor,
-    storage_context,
-)
 from galaxy.model.base import transaction
 from galaxy.model.dataset_collections import builder
 from galaxy.model.dataset_collections.matching import MatchingCollections
@@ -47,6 +43,10 @@ from galaxy.model.mapping import GalaxyModelMapping
 from galaxy.schema.schema import DatasetCollectionInstanceType
 from galaxy.schema.tasks import PrepareDatasetCollectionDownload
 from galaxy.security.idencoding import IdEncodingHelper
+from galaxy.short_term_storage import (
+    ShortTermStorageMonitor,
+    storage_context,
+)
 from galaxy.util import validation
 
 log = logging.getLogger(__name__)
