+++ lib/galaxy/managers/markdown_util.py
@@ -46,14 +46,14 @@ from galaxy.managers.jobs import (
     summarize_job_parameters,
 )
 from galaxy.managers.licenses import LicensesManager
-from galaxy.managers.short_term_storage import (
-    ShortTermStorageMonitor,
-    storage_context,
-)
 from galaxy.model.item_attrs import get_item_annotation_str
 from galaxy.model.orm.now import now
 from galaxy.schema import PdfDocumentType
 from galaxy.schema.tasks import GeneratePdfDownload
+from galaxy.short_term_storage import (
+    ShortTermStorageMonitor,
+    storage_context,
+)
 from galaxy.util.markdown import literal_via_fence
 from galaxy.util.resources import resource_string
 from galaxy.util.sanitize_html import sanitize_html
